+++ mongo/request_manager.go
@@ -609,15 +609,6 @@ func (r *RequestManager) RevokeRefreshToken(ctx context.Context, requestID strin
 		return err
 	}
 
-	err = r.Cache.Delete(ctx, storage.EntityCacheRefreshTokens, cacheObject.Key())
-	if err != nil {
-		// Log to StdOut
-		log.WithError(err).Error(logError)
-		// Log to OpenTracing
-		otLogErr(span, err)
-		return err
-	}
-
 	return nil
 }
 
@@ -669,15 +660,6 @@ func (r *RequestManager) RevokeAccessToken(ctx context.Context, requestID string
 	}
 
 	err = r.DeleteAccessTokenSession(ctx, cacheObject.Value())
-	if err != nil {
-		// Log to StdOut
-		log.WithError(err).Error(logError)
-		// Log to OpenTracing
-		otLogErr(span, err)
-		return err
-	}
-
-	err = r.Cache.Delete(ctx, storage.EntityCacheAccessTokens, cacheObject.Key())
 	if err != nil {
 		// Log to StdOut
 		log.WithError(err).Error(logError)