+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [v0.18.9] - 2020-06-13
+### Fixed
+- mongo: `RevokeAccessToken` attempted to delete the access token twice from 
+  the datastore leading to `fosite.ErrNotFound` always being returned.
+- mongo: `RevokeRefreshToken` attempted to delete the refresh token twice from 
+  the datastore leading to `fosite.ErrNotFound` always being returned.
+
 ## [v0.18.8] - 2020-06-11
 ### Fixed
 - mongo: auth codes should be set to active by default on creation.
@@ -451,6 +458,7 @@ clear out the password field before sending the response.
 - General pre-release!
 
 [Unreleased]: https://github.com/matthewhartstonge/storage/tree/master
+[v0.18.9]: https://github.com/matthewhartstonge/storage/tree/v0.18.9
 [v0.18.8]: https://github.com/matthewhartstonge/storage/tree/v0.18.8
 [v0.18.7]: https://github.com/matthewhartstonge/storage/tree/v0.18.7
 [v0.18.6]: https://github.com/matthewhartstonge/storage/tree/v0.18.6